GerichtHinzufuegenActivity.kt    new -> new add ingredient button
GerichtHinzufuegenActivity.kt
   new -> new ingredient listview

+++ app/src/main/java/de/pascalkuehnold/essensplaner/MainActivity.kt
@@ -1,5 +1,6 @@
 package de.pascalkuehnold.essensplaner
 
+import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
@@ -7,6 +8,7 @@ import android.view.Menu
 import android.view.MenuItem
 import android.widget.Button
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import de.pascalkuehnold.essensplaner.activities.Wochenplaner
 import de.pascalkuehnold.essensplaner.activities.GerichteListeActivity
 import de.pascalkuehnold.essensplaner.database.AppDatabase
@@ -34,8 +36,6 @@ class MainActivity : AppCompatActivity() {
 
     }
 
-
-
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         // Inflate the menu; this adds items to the action bar if it is present.
         menuInflater.inflate(R.menu.menu_main, menu)
@@ -58,12 +58,24 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
+    //Method for deleting the content on the phone (database of meals)
     private fun deleteDatabase(){
         println("Datenbank löschen gedrückt")
 
-        AppDatabase.getDatabase(applicationContext).gerichtDao().delete()
 
-        Toast.makeText(this, "Daten wurden gelöscht...", Toast.LENGTH_LONG).show()
+        val alert = AlertDialog.Builder(this)
+        alert.setTitle(getString(R.string.deleteDataFromPhone))
+        alert.setMessage(getString(R.string.deleteDataFromPhoneMessage))
+        alert.setPositiveButton(getString(R.string.delete)) { _: DialogInterface, _: Int ->
+            AppDatabase.getDatabase(applicationContext).gerichtDao().delete()
+            Toast.makeText(this, getString(R.string.allDataDeletedText), Toast.LENGTH_LONG).show()
+        }
+        alert.setNegativeButton(getString(R.string.cancel)) { dialog: DialogInterface, _: Int ->
+            dialog.cancel()
+        }
+        alert.show()
+
+
 
         println("Datenbank wurde gelöscht")
     }
